refactor semantic kitti dataloader to segmentation directory
